Connects to API and loads data into the list, todo: markers and filters
