+++ NomadMapApp/composeApp/src/androidMain/kotlin/org/nomad/mapapp/MainActivity.kt
@@ -46,7 +46,6 @@ class MainActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         // Request location permission
         when {
             ContextCompat.checkSelfPermission(
