+++ NomadMapApp/composeApp/src/androidMain/kotlin/org/nomad/mapapp/ui/component/FilterDialog.kt
@@ -19,16 +19,7 @@ fun FilterDialog(
 ) {
     val availableTags = remember {
         listOf(
-            "Todos", "accesorios", "Algodón Pima Orgánico", "alquiler de prendas y accesorios",
-            "Artesanía", "Bajo Impacto", "bolsos veganos", "Branding", "calcetines",
-            "diseño de estampados", "Diseño y arte", "Genderless", "haramaki", "marketplace",
-            "materia prima", "moda", "Moda Inclusiva", "Moda infantil", "Moda mujer",
-            "Moda sin género", "outlet", "Producción local", "Recién nacido",
-            "revalorización segunda mano", "Ropa de baño", "ropa de hombre", "ropa de mujer",
-            "ropa ecológica y ética", "ropa infantil", "ropa sostenible",
-            "Ropa sostenible pintada a mano", "Second Hand", "Segunda mano", "SOLIDARIA",
-            "textil hogar", "textil orgánico", "textiles reciclados sostenibles",
-            "tienda multimarca", "zapatos veganos"
+            "Ecològic", "Artesanal", "Local", "Moda ètica", "Zero residus", "Comerç just", "Reciclat", "Segona mà", "Sostenible", "Vegà"
         )
     }
